// src/types/productType.ts
import { PhotosType } from "./productPhoto";
import { ProductVariantsType } from "./productVariant";
import { SubCategoryType } from "./subCategoryType";

export type ProductType = {
    id: number;
    name:string;
    price: number;
    slug: string;
    description: string;
    thumbnail: string;
    is_ready: boolean;
    subCategory: SubCategoryType;
    productVariants: ProductVariantsType[];
    photos: PhotosType[];
}

// type for new arrivals & sub Categories carousel on Home Page
export type ProductThumbnailType = Pick<ProductType, 'id' | 'name' | 'thumbnail' | 'slug'>